[FIX] Logic and View separated
