+++ PokerApp/PokerApp/main.swift
@@ -11,7 +11,11 @@ let numOfUser = 4
 
 do {
     let poker: PokerPlay = try PokerPlay(type: PokerPlay.GameType.five, numOfUser: numOfUser)
-    poker.whoIsWinner()
+    if let winner = poker.whoIsWinner() {
+        let presenter: PokerPlayPresenter = PokerPlayPresenter(gamePlayer: poker.gamePlayer, winner: winner)
+        presenter.printResult()
+    }
+    
 } catch {
     print(error)
 }
